pyproject.toml
[tool.poetry]
name = "essensfindung"
version = "0.1.0"
description = "Student project for DHBW-Friedrichshafen to search for restaurant or recipe suggestions."
authors = ["Florian Glaser", "Florian Herkommer", "David Felder", "Baldur Siegel", "Jonas Straub"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.1"
uvicorn = "^0.16.0"
Jinja2 = "^3.0.3"
aiofiles = "^0.8.0"
pydantic = "^1.9.0"
httpx = "^0.21.3"
SQLAlchemy = "^1.4.31"
psycopg2 = "^2.9.3"
passlib = "^1.7.4"
bcrypt = "^3.2.0"
gunicorn = "^20.1.0"
python-dotenv = "^0.19.2"
python-multipart = "^0.0.5"
python-jose = "^3.3.0"
pandas = "^1.4.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
black = "^21.10b0"
flake8 = "^4.0.1"
pylint = "^2.11.1"
pytest = "^6.2.5"
pytest-httpx = "^0.17.3"
pytest-mock = "^3.6.1"
pre-commit = "^2.17.0"
sphinx-rtd-theme = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120